Review Pass 3 initial diff/changes
