+++ patches/tModLoader/Terraria/ModLoader/ModPylon.cs
@@ -1,6 +1,17 @@
 ﻿using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
+using ReLogic.Content;
+using System;
+using Terraria.Audio;
+using Terraria.DataStructures;
 using Terraria.GameContent;
+using Terraria.GameContent.ObjectInteractions;
+using Terraria.GameInput;
+using Terraria.ID;
+using Terraria.Localization;
 using Terraria.Map;
+using Terraria.ObjectData;
+using Terraria.UI;
 
 namespace Terraria.ModLoader
 {
@@ -180,5 +191,163 @@ namespace Terraria.ModLoader
 		/// <param name="deselectedScale"> The scale of the icon if it is NOT currently being hovered over. In vanilla, this is 1f, or 100%. </param>
 		/// <param name="selectedScale"> The scale of the icon if it IS currently being over. In vanilla, this is 2f, or 200%. </param>
 		public virtual void DrawMapIcon(ref MapOverlayDrawContext context, ref string mouseOverText, TeleportPylonInfo pylonInfo, bool isNearPylon, Color drawColor, float deselectedScale, float selectedScale) { }
+
+		public override bool RightClick(int i, int j) {
+			//Just like vanilla pylons, upon right click, open the map and play a sound.
+			Main.mapFullscreen = true;
+			SoundEngine.PlaySound(SoundID.MenuOpen);
+			return true;
+		}
+
+		//Must be true in order for our highlight texture to work.
+		public override bool HasSmartInteract(int i, int j, SmartInteractScanSettings settings) {
+			return true;
+		}
+
+		public override void DrawEffects(int i, int j, SpriteBatch spriteBatch, ref TileDrawInfo drawData) {
+			//Basically, we only want the crystal to be drawn once, based off the top left corner, which is what this check is.
+			//The top left corner of a Pylon will have its FrameX divisible by 54 (since it's 3 tiles wide, and each tile on the tile sheet takes up 18x18 pixels),
+			//and its FrameY will be 0, since it's at the top of the tile sheet.
+			if (drawData.tileFrameX % TileObjectData.GetTileData(drawData.tileCache).CoordinateFullWidth == 0 && drawData.tileFrameY == 0) {
+				//This method call basically says "Run SpecialDraw once at this position"
+				Main.instance.TilesRenderer.AddSpecialLegacyPoint(i, j);
+			}
+		}
+
+		/// <summary>
+		/// Draws the passed in pylon crystal texture the exact way that vanilla draws it. This MUST be called in SpecialDraw in order
+		/// to function properly.
+		/// </summary>
+		/// <param name="spriteBatch"> The sprite batch that will draw the crystal. </param>
+		/// <param name="i"> The X tile coordinate to start drawing from. </param>
+		/// <param name="j"> The Y tile coordinate to star drawing from. </param>
+		/// <param name="crystalTexture"> The texture of the crystal that will actually be drawn. </param>
+		/// <param name="crystalDrawColor"> The color to draw the crystal as. </param>
+		/// <param name="crystalHorizontalFrameCount">The total frames wide the crystal texture has. </param>
+		/// <param name="crystalVerticalFrameCount"> The total frames high the crystal texture has. </param>
+		public void DefaultDrawPylonCrystal(SpriteBatch spriteBatch, int i, int j, Asset<Texture2D> crystalTexture, Color crystalDrawColor, int crystalHorizontalFrameCount, int crystalVerticalFrameCount) {
+			//This is lighting-mode specific, always include this if you draw tiles manually
+			Vector2 offScreen = new Vector2(Main.offScreenRange);
+			if (Main.drawToScreen) {
+				offScreen = Vector2.Zero;
+			}
+
+			//Take the tile, check if it actually exists
+			Point pos = new Point(i, j);
+			Tile tile = Main.tile[pos.X, pos.Y];
+			if (tile == null || !tile.HasTile) {
+				return;
+			}
+
+			//Here, lots of framing takes place. 
+			Texture2D vanillaPylonCrystals = TextureAssets.Extra[181].Value; //The default textures for vanilla pylons, containing a "sheen" texture we need
+			int frameY = (Main.tileFrameCounter[TileID.TeleportationPylon] + pos.X + pos.Y) % 64 / crystalVerticalFrameCount; //Get the current frameY. All pylons share the same frameCounter
+			//Next, frame the actual crystal texture, it's highlight texture, and the sheen texture, in that order.
+			Rectangle crystalFrame = crystalTexture.Frame(crystalHorizontalFrameCount, crystalVerticalFrameCount, 0, frameY); 
+			Rectangle highlightFrame = crystalTexture.Frame(crystalHorizontalFrameCount, crystalVerticalFrameCount, 1, frameY);
+			vanillaPylonCrystals.Frame(crystalHorizontalFrameCount, crystalVerticalFrameCount, 0, frameY);
+			//Calculate positional values in order to determine where to actually draw the pylon
+			Vector2 origin = crystalFrame.Size() / 2f;
+			Vector2 centerPos = pos.ToWorldCoordinates(24f, 64f);
+			float centerDisplacement = (float)Math.Sin(Main.GlobalTimeWrappedHourly * ((float)Math.PI * 2f) / 5f);
+			Vector2 drawPos = centerPos + offScreen + new Vector2(0f, -40f) + new Vector2(0f, centerDisplacement * 4f);
+			//Randomly spawn dust, granted that the game is open an active
+			if (!Main.gamePaused && Main.instance.IsActive && Main.rand.NextBool(40)) {
+				Rectangle dustBox = Utils.CenteredRectangle(drawPos, crystalFrame.Size());
+
+				int dustIndex = Dust.NewDust(dustBox.TopLeft(), dustBox.Width, dustBox.Height, DustID.TintableDustLighted, 0f, 0f, 254, Color.Gray, 0.5f);
+				Main.dust[dustIndex].velocity *= 0.1f;
+				Main.dust[dustIndex].velocity.Y -= 0.2f;
+			}
+
+			//Next, calculate the color of the crystal and draw it. The color is determined by how lit the tile is at that position.
+			Color crystalColor = Color.Lerp(Lighting.GetColor(pos.X, pos.Y), crystalDrawColor, 0.8f) ;
+			spriteBatch.Draw(crystalTexture.Value, drawPos - Main.screenPosition, crystalFrame, crystalColor * 0.7f, 0f, origin, 1f, SpriteEffects.None, 0f);
+
+			//Next, calculate the color of the sheen texture and its scale, then draw it.
+			float scale = (float)Math.Sin(Main.GlobalTimeWrappedHourly * (Math.PI * 2f) / 1f) * 0.2f + 0.8f;
+			Color sheenColor = new Color(255, 255, 255, 0) * 0.1f * scale;
+			for (float displacement = 0f; displacement < 1f; displacement += 355f / (678f * (float)Math.PI)) {
+				spriteBatch.Draw(crystalTexture.Value, drawPos - Main.screenPosition + ((float)Math.PI * 2f * displacement).ToRotationVector2() * (6f + centerDisplacement * 2f), crystalFrame, sheenColor, 0f, origin, 1f, SpriteEffects.None, 0f);
+			}
+
+			//Finally, everything below is for the smart cursor, drawing the highlight texture depending on whether or not either:
+			//1) Smart Cursor is off, and no highlight texture is drawn at all (selectionType = 0)
+			//2) Smart Cursor is on, but it is not currently focusing on the Pylon (selectionType = 1)
+			//3) Smart Cursor is on, but it IS currently focusing on the Pylon (selectionType = 2)
+			int selectionType = 0;
+			if (Main.InSmartCursorHighlightArea(pos.X, pos.Y, out bool actuallySelected)) {
+				selectionType = 1;
+
+				if (actuallySelected) {
+					selectionType = 2;
+				}
+			}
+
+			if (selectionType == 0) {
+				return;
+			}
+
+			//Finally, draw the highlight texture, if applicable.
+			int colorPotency = (crystalColor.R + crystalColor.G + crystalColor.B) / 3;
+			if (colorPotency > 10) {
+				Color selectionGlowColor = Colors.GetSelectionGlowColor(selectionType == 2, colorPotency);
+				spriteBatch.Draw(crystalTexture.Value, drawPos - Main.screenPosition, highlightFrame, selectionGlowColor, 0f, origin, 1f, SpriteEffects.None, 0f);
+			}
+		}
+
+		/// <summary>
+		/// Draws the passed in map icon texture for pylons the exact way that vanilla would draw it. Note that this method
+		/// assumes that the texture is NOT framed, i.e there is only a single sprite that is not animated.
+		/// Returns whether or not the player is currently hovering over the icon.
+		/// </summary>
+		/// <param name="context"> The draw context that will allow for drawing on thj</param>
+		/// <param name="mapIcon"> The icon that is to be drawn on the map. </param>
+		/// <param name="drawCenter"> The position in TILE coordinates for where the CENTER of the map icon should be. </param>
+		/// <param name="drawColor"> The color to draw the icon as. </param>
+		/// <param name="deselectedScale"> The scale to draw the map icon when it is not selected (not being hovered over) </param>
+		/// <param name="selectedScale"> The scale to draw the map icon when it IS selected (being hovered over)</param>
+		/// <returns></returns>
+		public bool DefaultDrawMapIcon(ref MapOverlayDrawContext context, Asset<Texture2D> mapIcon, Vector2 drawCenter, Color drawColor, float deselectedScale, float selectedScale) {
+			return context.Draw(
+				              mapIcon.Value,
+				              drawCenter,
+				              drawColor,
+				              new SpriteFrame(1, 1, 0, 0),
+				              deselectedScale,
+				              selectedScale,
+				              Alignment.Center
+				              )
+			              .IsMouseOver;
+		}
+
+		/// <summary>
+		/// Handles mouse clicking on the map icon the exact way that vanilla handles it. In normal circumstances, this should be called
+		/// directly after DefaultDrawMapIcon.
+		/// </summary>
+		/// <param name="mouseIsHovering"> Whether or not the map icon is currently being hovered over. </param>
+		/// <param name="pylonInfo"> The information pertaining to the current pylon being drawn. </param>
+		/// <param name="hoveringTextKey">
+		/// The localization key that will be used to display text on the mouse, granted the mouse is currently hovering over the map icon.
+		/// </param>
+		/// <param name="mouseOverText"> The reference to the string value that actually changes the mouse text value. </param>
+		public void DefaultMapClickHandle(bool mouseIsHovering, TeleportPylonInfo pylonInfo, string hoveringTextKey, ref string mouseOverText) {
+			//We only want these things to happen if the mouse is hovering, thus the check:
+			if (!mouseIsHovering) {
+				return;
+			}
+
+			Main.cancelWormHole = true;
+			mouseOverText = Language.GetTextValue(hoveringTextKey);
+
+			//If clicking, then teleport!
+			if (Main.mouseLeft && Main.mouseLeftRelease) {
+				Main.mouseLeftRelease = false;
+				Main.mapFullscreen = false;
+				PlayerInput.LockGamepadButtons("MouseLeft");
+				Main.PylonSystem.RequestTeleportation(pylonInfo, Main.LocalPlayer);
+				SoundEngine.PlaySound(SoundID.Item6);
+			}
+		}
 	}
 }
